backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
supabase>=2.10.0
alembic==1.13.1

# Redis (para sessões)
redis==5.0.1
hiredis==2.3.2

# AI / OpenAI
openai>=1.30.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# PDF
PyPDF2>=3.0.0

# Logging
loguru==0.7.2

# Utils
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic[email]==2.5.3
PyYAML==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
requests==2.31.0
